Avatar size problem solved
DONE:
Rework the SelectedImage function to be able to check the file type and size.
Put an error message if invalid file.
The Update button is disabled until a vild file is selected.
When a valid file is selected, the button is enabled and the color is changed to NeonBlue.
Work on Avatar_Modal.js / Tools.js / Avatar.html